# api-docs-EntidadBancaria.json
{"openapi":"3.1.0","info":{"title":"Mi API","description":"Descripción de mi API","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/bancos/actualizar/{id}":{"put":{"tags":["entidad-bancaria-ctrl"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntidadBancaria"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntidadBancaria"}}}}}}},"/api/bancos/registrar":{"post":{"tags":["entidad-bancaria-ctrl"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntidadBancariaDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/cliente/self/{id}":{"get":{"tags":["client-self-ctrl"],"operationId":"ObtenerBancoComoCliente","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntidadBancaria"}}}}}}},"/api/bancos/obtener/{id}":{"get":{"tags":["entidad-bancaria-ctrl"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntidadBancaria"}}}}}}},"/api/bancos/listar":{"get":{"tags":["entidad-bancaria-ctrl"],"operationId":"findAllBanco","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EntidadBancaria"}}}}}}}},"/api/bancos/eliminar/{id}":{"delete":{"tags":["entidad-bancaria-ctrl"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"EntidadBancaria":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"codigoEntidad":{"type":"string"},"nombre":{"type":"string"},"cuit":{"type":"string"},"direccion":{"type":"string","maxLength":200,"minLength":0},"tipoEntidad":{"type":"string","enum":["BANCO","CAJA_AHORRO","COOPERATIVA","FINANCIERA","CASA_CAMBIO","OTRO"]},"activa":{"type":"boolean"},"fechaCreacion":{"type":"string","format":"date-time"}}},"EntidadBancariaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"codigoEntidad":{"type":"string","maxLength":5,"minLength":3},"nombre":{"type":"string","maxLength":100,"minLength":0},"cuit":{"type":"string","minLength":1,"pattern":"\\d{11}"},"direccion":{"type":"string","minLength":1},"tipoEntidad":{"type":"string","enum":["BANCO","CAJA_AHORRO","COOPERATIVA","FINANCIERA","CASA_CAMBIO","OTRO"]},"activa":{"type":"boolean"}},"required":["codigoEntidad","cuit","direccion","nombre","tipoEntidad"]}}}}